+++ ext/json/php_json.h
@@ -86,8 +86,8 @@ ZEND_BEGIN_MODULE_GLOBALS(json)
 	int encoder_depth;
 	int encode_max_depth;
 	php_json_error_code error_code;
-	int error_line;
-	int error_column;
+	size_t error_line;
+	size_t error_column;
 ZEND_END_MODULE_GLOBALS(json)
 
 PHP_JSON_API ZEND_EXTERN_MODULE_GLOBALS(json)
